requirements.txt
Flask==3.1.0
ebooklib==0.18
Pillow==11.1.0
psutil==6.1.1
python-dotenv==1.0.1
waitress==3.0.2
